requirements.txt
absl-py==2.1.0
aiolimiter==1.1.0
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Deprecated==1.2.14
exceptiongroup==1.2.1
filelock==3.13.4
fsspec==2024.3.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.22.2
ibm-generative-ai==3.0.0
idna==3.6
Jinja2==3.1.3
joblib==1.4.0
jsonpatch==1.33
jsonpointer==2.4
langchain-core==0.1.52
langsmith==0.1.63
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
orjson==3.10.3
packaging==23.2
pillow==10.3.0
pydantic==2.7.2
pydantic_core==2.18.3
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rouge-score==0.1.2
safetensors==0.4.2
scikit-learn==1.4.2
scipy==1.13.0
sentence-transformers==2.6.1
six==1.16.0
sniffio==1.3.1
sympy==1.12
tenacity==8.3.0
threadpoolctl==3.4.0
tokenizers==0.15.2
torch==2.2.2
tqdm==4.66.2
transformers==4.39.3
typing_extensions==4.11.0
urllib3==2.2.1
wrapt==1.16.0
